+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/LoopVideoPatch.java
@@ -1,0 +1,13 @@
+package app.revanced.extension.youtube.patches;
+
+import app.revanced.extension.youtube.settings.Settings;
+
+@SuppressWarnings("unused")
+public class LoopVideoPatch {
+    /**
+     * Injection point
+     */
+    public static boolean shouldLoopVideo() {
+        return Settings.LOOP_VIDEO.get();
+    }
+}