+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/buttons/action/HideButtonsPatch.kt
@@ -23,7 +23,6 @@ val hideButtonsPatch = resourcePatch(
     compatibleWith(
         "com.google.android.youtube"(
             "19.34.42",
-            "19.43.41",
             "20.07.39",
             "20.13.41",
             "20.14.43",