+++ patches/src/main/kotlin/app/revanced/patches/youtube/layout/hide/relatedvideooverlay/HideRelatedVideoOverlayPatch.kt
@@ -49,7 +49,6 @@ val hideRelatedVideoOverlayPatch = bytecodePatch(
     compatibleWith(
         "com.google.android.youtube"(
             "19.34.42",
-            "19.43.41",
             "20.07.39",
             "20.13.41",
             "20.14.43",